Pressure Sensor Code Integrated + Arduino Unit Test Start
Implemented controller module for getting pressure readings from the 3 pressure sensors used in the ventilator design as per the HW ICD. Wrote basic unit test.

Wrote first steps for Arduino unit testing. Created fake Arduino wrapper so unit tests can be run independently of the Arduino. This can be reused and added to for further unit testing of other modules.

Module succesfully passed unit test. Can use module as part of a simulation for the master state machine without needing to run it on an Arduino.