Closes Issue #112
platformio test -e native now passes with SensorTests FullScaleRead enabled.

Error was due to the original sensor module being refactored to use the new HAL API but the sensortests portion was not updated and still used my old HAL. Thus it broke due to the split usage. I refactored the test to use the new HAL API.

Removed my Arduino HAL which is replaced by the new HAL and GMock (which are much better).

Removed the sensor tests header because it's not needed under the new testing framework.