Add alg.h, which defines non-macro versions of min/max.
Since we don't have <algorithm>, define a "polyfill" for min/max.  This
is particularly important for code where we might want to write
something like:

  readIntoBuffer(buf, min(bufLen, bytesReady()))

If `min` is interpreted as the macro `a < b ? a : b`, bytesReady() may
be called twice.  The first time it's called, maybe its return value is
less than bufLen, and the second time, who knows, maybe its value is
greater than bufLen.  In that case, we have a buffer overflow.

The functional version doesn't have this bug.