Move serialIO into HAL.
This makes serialIO testable.

I found but did not fix a major issue in data transmission.  Namely, the
serial link's outgoing buffer is a fixed size (something like 64 bytes).
Before this change, the controller would block if that filled up -- in
other words, if the RPi crashes, the controller hangs.  After my change,
the situation is only slightly better: Now if the RPi doesn't read from
our transmit buffer, the controller will silently drop data.  This is
still Very Bad because we have no way to re-synchronize the stream after
dropping anything.

The HAL code itself is *just* complex enough that it might be worth
having tests.  OTOH it's test-only code, and presumably tests for other
components will also test it, so maybe that's not the end of the world.