Simplify sensor calibration routine.
The zero_sensors() function is unsafe to call; sleeping for 100ms can
trip the watchdog (currently set to a 120ms timeout).  In fact the old
code was doing a 120ms sleep, so was coming *extremely close* to
tripping the watchdog.

I've deleted the zero_sensors() function (we now can calibrate only
during startup) and reduced the amount of time we sleep.

It happens that this is *still* unsafe.  Calibration has to happen while
the device is quiescent, but we *can't* safely wait until quiescence if
we're restarted while we're ventilating a patient.

It seems that we need to save calibration data to non-volatile memory
and then give operators an option to recalibrate when a patient is not
hooked up to the device.

Similarly I've deleted the scaffold blower module, since this seems very
challenging to use safely.

There's important work to be done to improve this, but I don't think the
approach we have scaffolded will work, so ripping it out is safer and
lets us carry less complexity.

git-pr-chain: sensor-init