Clarify that we're using MPXV5004DP pressure sensors everywhere.
Previously we were using the V7002 for some of our sensors.

No code changes are needed -- and we can actually test this! -- because
the only difference between the two from our perspective is that they
give different voltages for zero pressure.  Our code dynamically finds
the zero value *anyway*, so it doesn't make a difference.

I'm still looking for the canonical place that will list exactly what
parts we're using, etc.  We'll be able to link to that from HAL once we
have it.

git-pr-chain: MPXV5004DP