Clarify that HAL is not a mock.
When I wrote it, I was anticipating it would be half-mock, half-fake.
We seem to be getting along fine with all-fake no-mock, so clarify the
comments.

Also link to the canonical list of parts, and get rid of unnecessary
enum int types (the compiler will chose them for us, this is an advanced
feature really only needed for cross-platform ABI compatibility and such).

git-pr-chain: hal-mock