Re-tune PID.
Edwin added an output constriction to the device -- i.e. he reduced the
diameter of the exhaust tube.  We then needed to re-tune PID.

We also needed to re-tune because my previous experiments were done at a
high PIP and PEEP, and this was causing the blower to max out.

GPC: retune-pid