Add BlockInterrupts RAII class.
This is a safer C++/RAII way to block and restore interrupts.

This slightly expands the critical section in FullCt().  I could reduce
it if that's a problem.

Also renames IntEnable/IntDisable to enableInterrupts() and
disableInterrupts().  These names make it so that it's unnecessary to
have comments explaining what the functions do.  Also our method names
start with lower case letters (apparently).

GPC: interrupts-raii