Added get and set commands
This updates the console debug utility to add the ability to get and set
debug variables.  Debug variables can be very easily added to the
controller for debugging purposes, no need to change the python utility.
These variables all parameters to be modified in real time.

This change also adds some debug variables for the controller tuning
gains, partly because that will be very handy during tuning, and partly
as an example of how these work