[GUI] Add initial implementation on Selection Mode Popup
Mode selection popup will allow us to change the ventilator
operation mode. The design and copies are still
under development and will be refined later.

The organization of the QML files was also changed for easy understanding:

* control/ folder keeps all Pure UI controls, like buttons and widgets
* modes/ folder keeps the implementation of all modes. Later this could be
broken down if it gets too complex.

Add stub modes: PressureAssistMode and SupportPressureMode for
testing the mode selection