[GUI] Fix run_eglfs.sh
The bash script environment variables were not properly configured.

Also depending on the screen you plug your device, the mode selected
in kms.json might not be available, causing the app to segfault.

We are commenting the variable that forces the specific mode, letting
Qt decide with mode to use.
The bash script environment variables were not properly configure.