controller_debug: Print help for trace.
When you run plain `trace`, print an error and also print out `trace
help`.

To accomplish this, I reformatted the do_trace docstring so there isn't
a bunch of leading whitespace.  This matches other docstrings in here.
I didn't change the content of the docstring in this change.

GPC: print-trace-help