[GUI] Update design of the graphs
Update the design of the graphs to follow new recommendations.

The current implementation of the ScopeView has a very high CPU usage,
making it unable to hold updates at 60Hz. (Actually I can at 90%
CPU usage).

The main reason is the fact we are using two Charts, one To draw a line,
and the other to draw the area. Hopefully, with enough time we can
customize the graphs to make it more efficient.