[GUI] Improve chart performance to run on 60Hz.
We found out[1] that at 15Hz drawing the charts would
put pressure on the CPU, making the app use 20-30% of CPU
time on a MacBook pro.

After some investigation, we came to the conclusion that
AreaSeries were rendering the Graph on a QImage (on CPU),
then turned into the texture.

We researched alternatives and decided to drop QtCharts and
use a high-performance canvas-style library called QNanoPainter.

So far the graphs look identical, holding CPU time of around 10%
on 60Hz on MacBook Pro and 25-30% on 60Hz Raspberry pi 4.

1 - https://github.com/RespiraWorks/VentilatorSoftware/pull/464#pullrequestreview-426754871